Add sample database table
Added method to check database to see if a case already exists. If it does, program will skip to the next row in the csv. Also created interpreter.isolateSamples(), which is able to identify individual samples and split up a cell string accordingly. The same functionality will be used to identify individual events.