Add simple Test interpretation, construction and insertion methods
Added methods for program to identify, create, and insert Tests for simple Cases. The concept of a Test as a specific examination within a pregnancy is poorly defined in the source spreadsheet, but an important distinction to make. One Case (pregnancy) will always have the same fetus, set of potential fathers, and one definitive result throughout the pregnancy. However, a Case can still have multiple Test run on it, each with a different cost, turnaround, gestation, etc. This update is meant to identify Tests from simple Cases (one Test per Case).

Signed-off-by: Ellis Rucker <m.ellisrucker@gmail.com>