Add Complex Case Filter with associated DDL and DML
Added ability to detect extreme cases and add them to separate table for later review. The source spreadsheet contains too many variables to account for and program against. The goal will be to have the program automatically parse and enter ~80% of cases into the database. The more extreme cases can be entered in manually. The case filter runs before the main part of the program, and can be adjusted as program becomes able to handle some of the more extreme cases.

Signed-off-by: Ellis Rucker <m.ellisrucker@gmail.com>