Add Updater.updateSamplesAndPatients() and related methods
Added ability to update Samples and Patients within the database. Simply deleting and replacing all samples in a test would result in the date of receipt being lost. Instead, the Samples that no longer appear in the cell are deleted, the patientID of Samples that still appear in the cell are updated, and then new Samples are inserted. Will look into a simpler solution after program is completed.

Remember to add ability to parse samples from 2nd and 3rd draw cells.

Signed-off-by: Ellis Rucker <m.ellisrucker@gmail.com>