Remove EventString class Add Updater.updateEvents
Added ability to update Genotypes and Plasmas within the database. Temporarily, all events are deleted and re-inserted if either one of genotypeA, genotypeB, first draw, second draw, or third draw cells are changed. But separate methods will later be added to allow for each cell to be updated independently. Like the SampleString class, the EventString class was removed to simplify the parsing process.

Signed-off-by: Ellis Rucker <m.ellisrucker@gmail.com>