Refactor ChangeMap Class into DTO Log Class
As part of goal to remove IntermediateObject package, the ChangeMap class was refactored into the Log Class and moved to the DataTransferObject package. The class was also expanded to support a new Log table in the database. These changes allow the program to document every instance a case is updated and log the date and filename associated with the update. The logging function will aid with debugging any unexpected behaviors within the program when migrating the patient data.

Signed-off-by: Ellis Rucker <m.ellisrucker@gmail.com>