Refactor database connection handling
Optimized code by reducing the frequency in which database connections are opened and closed during runtime. Data manipulation language in DTO classes were moved to higher level classes to facilitate this. Turned off auto-commit for connections to allow for the database to be rolled back to its previous state in the event of an error.

Signed-off-by: Ellis Rucker <m.ellisrucker@gmail.com>